+++ dev_test_dex.py
@@ -41,7 +41,7 @@ import os
 
 
 # https://docs.python.org/3/library/logging.html#logging-levels
-logging.basicConfig(level=logging.ERROR,
+logging.basicConfig(level=logging.INFO,
                     filename=os.path.basename(__file__) + '.log',
                     format="{asctime} [{levelname:8}] {process} {thread} {module}: {message}",
                     style="{")
@@ -60,17 +60,14 @@ def print_stream_data_from_stream_buffer(binance_websocket_api_manager):
 
 
 # create instance of BinanceWebSocketApiManager for Binance Chain DEX
-binance_websocket_api_manager = BinanceWebSocketApiManager(exchange="binance.org")
-
-print("starting monitoring api!")
-binance_websocket_api_manager.start_monitoring_api()
+binance_websocket_api_manager = BinanceWebSocketApiManager(exchange="binance.org-testnet")
 
 # start a worker process to move the received stream_data from the stream_buffer to a print function
 worker_thread = threading.Thread(target=print_stream_data_from_stream_buffer, args=(binance_websocket_api_manager,))
 worker_thread.start()
 
 # userAddress streams
-binance_dex_user_address = ""
+binance_dex_user_address = "bnb1v566f3avl2ud5z0jepazsrguzkj367snlx4jm6"
 id = binance_websocket_api_manager.create_stream('orders', binance_dex_user_address)
 binance_websocket_api_manager.create_stream('accounts', binance_dex_user_address)
 binance_websocket_api_manager.create_stream('transfers', binance_dex_user_address)
@@ -81,7 +78,7 @@ binance_websocket_api_manager.create_stream(["allTickers"], ["$all"])
 binance_websocket_api_manager.create_stream(["allMiniTickers"], ["$all"])
 binance_websocket_api_manager.create_stream(["blockheight"], ["$all"])
 
-markets = 'RAVEN-F66_BNB'
+markets = ['0KI-0AF_BNB', '000-0E1_BNB', '1KVOLUME-D65_BNB', '81JIAN-3E8_BNB', 'AAA-B50_BNB', 'AAAAAA-BBA_BNB']
 binance_websocket_api_manager.create_stream(["trades"], markets)
 binance_websocket_api_manager.create_stream(["marketDepth"], markets)
 binance_websocket_api_manager.create_stream(["kline_1m"], markets)
@@ -89,10 +86,7 @@ binance_websocket_api_manager.create_stream(["kline_1h"], markets)
 binance_websocket_api_manager.create_stream(["ticker"], markets)
 binance_websocket_api_manager.create_stream(["miniTicker"], markets)
 
-markets = ['RAVEN-F66_BNB', 'ANKR-E97_BNB', 'AWC-986_BNB', 'COVA-218_BNB', 'BCPT-95A_BNB', 'WISH-2D5_BNB',
-           'MITH-C76_BNB', 'BNB_BTCB-1DE', 'BNB_USDSB-1AC', 'BTCB-1DE_USDSB-1AC', 'NEXO-A84_BNB']
-channels = ['trades']
-channels = ['kline_1m', 'kline_5m', 'kline_15m', 'marketDepth', 'ticker', 'miniTicker', 'marketDiff', 'trades']
+channels = ['kline_1m', 'kline_5m', 'kline_15m', 'marketDepth', 'ticker', 'miniTicker', 'marketDiff', 'trades', 'allTickers']
 multiplex_stream_id = binance_websocket_api_manager.create_stream(channels, markets)
 if binance_websocket_api_manager.wait_till_stream_has_started(multiplex_stream_id):
     binance_websocket_api_manager.print_stream_info(multiplex_stream_id)
@@ -100,11 +94,11 @@ if binance_websocket_api_manager.wait_till_stream_has_started(multiplex_stream_i
 
     binance_websocket_api_manager.subscribe_to_stream(multiplex_stream_id,
                                                       channels=['kline_1m', 'kline_5m', 'kline_15m', 'marketDepth',
-                                                                'ticker', 'miniTicker', 'marketDiff'])
+                                                                'ticker', 'marketDiff'])
 
-stream_id = binance_websocket_api_manager.create_stream(["kline_1m"], ['RAVEN-F66_BNB'])
+stream_id = binance_websocket_api_manager.create_stream(["kline_1m"], markets)
 
 while True:
-    binance_websocket_api_manager.print_summary()
-    #binance_websocket_api_manager.print_stream_info(stream_id)
+    #binance_websocket_api_manager.print_summary()
+    binance_websocket_api_manager.print_stream_info(multiplex_stream_id)
     time.sleep(1)